Tests/LibWeb/Text/expected/DOM/Document-dir.txt

String
document.dir = ''
document.dir = 'ltr'
document.dir = 'rtl'
document.dir = 'auto'
document.dir = ''
